requirements.txt
bottle ~= 0.12.19
pyotp ~= 2.6.0
PyYAML ~= 6.0.0